src/api/requirements.txt
affine==2.4.0
alembic==1.10.3
anyio==3.6.2
attrs==23.1.0
autopep8==2.0.2
azure-core==1.27.1
azure-storage-blob==12.9.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
click-plugins==1.1.1
cligj==0.7.2
cryptography==41.0.1
databases==0.7.0
fastapi==0.95.0
fastapi-pagination==0.9.1
Fiona==1.8.22
greenlet==2.0.2
gunicorn==20.1.0
h11==0.14.0
idna==3.4
importlib-metadata==6.6.0
importlib-resources==5.12.0
isodate==0.6.1
Mako==1.2.4
MarkupSafe==2.1.2
msrest==0.7.1
munch==2.5.0
numpy==1.24.3
oauthlib==3.2.2
psycopg2-binary==2.9.6
pycodestyle==2.10.0
pycparser==2.21
pydantic==1.10.7
pyparsing==3.0.9
python-dotenv==1.0.0
python-multipart==0.0.6
rasterio==1.3.6
rasterstats==0.18.0
requests==2.31.0
requests-oauthlib==1.3.1
shapely==2.0.1
simplejson==3.19.1
six==1.16.0
sniffio==1.3.0
snuggs==1.4.7
SQLAlchemy==1.4.47
starlette==0.26.1
tomli==2.0.1
typing-extensions==4.5.0
urllib3==2.0.3
uvicorn==0.21.1
zipp==3.15.0